Implement `typeFullNameForLiteral` and `typeFullNameForIdentPat` via rust_ast_gen.
