+++ joern-cli/frontends/rust2cpg/src/main/scala/io/joern/rust2cpg/astcreation/RustFullNames.scala
@@ -32,4 +32,26 @@ trait RustFullNames { this: AstCreator =>
       .orElse(path.pathSegment.nameRef.flatMap(rustAstGenTypeFullName))
       .getOrElse(Defines.Any)
   }
+  
+  protected def typeFullNameForLiteral(lit: RustNodeSyntax.Literal): String =
+    rustAstGenTypeFullName(lit)
+      .orElse(lit.value.map(typeFullNameForLiteralToken))
+      .getOrElse(Defines.Any)
+  
+  protected def typeFullNameForIdentPat(identPat: RustNodeSyntax.IdentPat): String = {
+    rustAstGenTypeFullName(identPat).getOrElse(Defines.Any)
+  }
+
+  protected def typeFullNameForLiteralToken(tok: RustNodeSyntax.RustToken): String = tok match {
+    case _: RustNodeSyntax.IntNumberToken   => "i32"
+    case _: RustNodeSyntax.FloatNumberToken => "f64"
+    case _: RustNodeSyntax.StringToken      => "&str"
+    case _: RustNodeSyntax.ByteStringToken  => "&[u8]"
+    case _: RustNodeSyntax.CStringToken     => "&CStr"
+    case _: RustNodeSyntax.CharToken        => "char"
+    case _: RustNodeSyntax.ByteToken        => "u8"
+    case _: RustNodeSyntax.TrueKwToken      => "bool"
+    case _: RustNodeSyntax.FalseKwToken     => "bool"
+    case _                                  => Defines.Any
+  }
 }
